Refactor envvar middleware config
